% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.sage_estimate}
\alias{print.sage_estimate}
\title{Print a SAGE Estimate}
\usage{
\method{print}{sage_estimate}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{sage_estimate}, as returned by
\code{\link[=sage_cv_gcop]{sage_cv_gcop()}}, \code{\link[=sage_cv]{sage_cv()}} or
\code{\link[=sage_cv_marginal]{sage_cv_marginal()}}.}

\item{digits}{Integer. Significant digits for the numeric columns.}

\item{...}{Passed to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
\code{x}, invisibly. Called for the side effect of printing.
}
\description{
Print a SAGE Estimate
}
\examples{
set.seed(1)
X <- matrix(rnorm(200 * 3), 200, 3)
colnames(X) <- paste0("X", 1:3)
res <- sage_cv_gcop(function(Xm) 2 * Xm[, 1], X, K = 2, M = 5, B = 6, seed = 1)
print(res)

}
